Rename stack status enum. Simplify code
